+++ src/CausalHorseForest.cpp
@@ -9,8 +9,11 @@ Rcpp::List CausalHorseForest_cpp(
   SEXP X_train_treatSEXP,
   SEXP X_train_controlSEXP, 
   SEXP ySEXP, 
-  SEXP status_indicatorSEXP, 
+  SEXP status_indicatorSEXP,
   SEXP is_survivalSEXP,
+  SEXP observed_left_timeSEXP,
+  SEXP observed_right_timeSEXP,
+  SEXP interval_censoring_indicatorSEXP,
   SEXP treatment_indicatorSEXP,
   SEXP n_testSEXP, 
   SEXP X_test_controlSEXP, 
@@ -70,11 +73,19 @@ Rcpp::List CausalHorseForest_cpp(
   double* y = &y_vector[0];  
   bool is_survival = Rcpp::as<bool>(is_survivalSEXP);
   Rcpp::IntegerVector treatment_indicator_vector(treatment_indicatorSEXP);
-  int* treatment_indicator = &treatment_indicator_vector[0];  
-  Rcpp::NumericVector status_indicator_vector(status_indicatorSEXP);   
-  double* status_indicator = &status_indicator_vector[0];   
+  int* treatment_indicator = &treatment_indicator_vector[0];
+  Rcpp::NumericVector status_indicator_vector(status_indicatorSEXP);
+  double* status_indicator = &status_indicator_vector[0];
   std::vector<double> y_observed_vector(y_vector.begin(), y_vector.end());  // Create an independent copy of y
   double* y_observed = y_observed_vector.data(); // Get pointer to use like an array
+
+  // Objects for interval censoring
+  Rcpp::NumericVector observed_left_time_vector(observed_left_timeSEXP);
+  double* observed_left_time = &observed_left_time_vector[0];
+  Rcpp::NumericVector observed_right_time_vector(observed_right_timeSEXP);
+  double* observed_right_time = &observed_right_time_vector[0];
+  Rcpp::NumericVector interval_censoring_indicator_vector(interval_censoring_indicatorSEXP);
+  double* interval_censoring_indicator = &interval_censoring_indicator_vector[0];
 
   // Parameters and data for the test phase
   size_t n_test = Rcpp::as<size_t>(n_testSEXP);    
@@ -475,15 +486,17 @@ Rcpp::List CausalHorseForest_cpp(
       random
     );
 
-    // Augment the censored data (outer Gibbs step), if applicable
+    // Augment the censored data (right-censored and interval-censored)
     AugmentCensoredObservations(
-      is_survival, 
-      y, 
-      y_observed, 
-      status_indicator, 
-      total_predictions, 
-      sigma, 
-      n, 
+      is_survival,
+      y,
+      observed_left_time,
+      status_indicator,
+      observed_right_time,
+      interval_censoring_indicator,
+      total_predictions,
+      sigma,
+      n,
       random
     );